+++ JVM/runtime/LocalVariablesArray.h
@@ -5,6 +5,6 @@ class LocalVariablesArray : public Array<word>
 {
 public:
 	LocalVariablesArray() : LocalVariablesArray(10) {};
-	LocalVariablesArray(int size) : Array(size) {};
+	LocalVariablesArray(size_t size) : Array(size) {};
 	~LocalVariablesArray() {}
 };
