+++ JVM/runtime/MethodFrame.cpp
@@ -6,7 +6,7 @@ MethodFrame::MethodFrame()
 {
 }
 
-MethodFrame::MethodFrame(int stackSize, int localVariablesSize): operandStack(stackSize), localVariables(localVariablesSize)
+MethodFrame::MethodFrame(size_t stackSize, size_t localVariablesSize): operandStack(stackSize), localVariables(localVariablesSize)
 {
 }
